+++ pkg/networkserver/adr_test.go
@@ -23,7 +23,7 @@ import (
 
 	pbtypes "github.com/gogo/protobuf/types"
 	"github.com/smartystreets/assertions"
-	"go.thethings.network/lorawan-stack/pkg/band"
+	"go.thethings.network/lorawan-stack/pkg/log"
 	"go.thethings.network/lorawan-stack/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/pkg/util/test"
 	"go.thethings.network/lorawan-stack/pkg/util/test/assertions/should"
@@ -230,26 +230,63 @@ func TestLossRate(t *testing.T) {
 }
 
 func TestAdaptDataRate(t *testing.T) {
+	semtechPaperUplinks := adrMatrixToUplinks([]adrMatrixRow{
+		{FCnt: 10, MaxSNR: -6, GtwDiversity: 2},
+		{FCnt: 11, MaxSNR: -7, GtwDiversity: 2},
+		{FCnt: 12, MaxSNR: -25, GtwDiversity: 1},
+		{FCnt: 13, MaxSNR: -25, GtwDiversity: 1},
+		{FCnt: 14, MaxSNR: -10, GtwDiversity: 2},
+		{FCnt: 16, MaxSNR: -25, GtwDiversity: 1},
+		{FCnt: 17, MaxSNR: -10, GtwDiversity: 2},
+		{FCnt: 19, MaxSNR: -10, GtwDiversity: 3},
+		{FCnt: 20, MaxSNR: -6, GtwDiversity: 2},
+		{FCnt: 21, MaxSNR: -7, GtwDiversity: 2},
+		{FCnt: 22, MaxSNR: -25, GtwDiversity: 0},
+		{FCnt: 23, MaxSNR: -25, GtwDiversity: 1},
+		{FCnt: 24, MaxSNR: -10, GtwDiversity: 2},
+		{FCnt: 25, MaxSNR: -10, GtwDiversity: 2},
+		{FCnt: 26, MaxSNR: -25, GtwDiversity: 1},
+		{FCnt: 27, MaxSNR: -8, GtwDiversity: 2},
+		{FCnt: 28, MaxSNR: -10, GtwDiversity: 2},
+		{FCnt: 29, MaxSNR: -10, GtwDiversity: 3},
+		{FCnt: 30, MaxSNR: -9, GtwDiversity: 3},
+		{
+			FCnt: 31, MaxSNR: -7, GtwDiversity: 2,
+			TxSettings: ttnpb.TxSettings{
+				DataRate: ttnpb.DataRate{
+					Modulation: &ttnpb.DataRate_LoRa{
+						LoRa: &ttnpb.LoRaDataRate{
+							SpreadingFactor: 12,
+							Bandwidth:       125000,
+						},
+					},
+				},
+				DataRateIndex: 0,
+			},
+		},
+	})
 	for _, tc := range []struct {
 		Name       string
 		Device     *ttnpb.EndDevice
-		PHY        band.Band
 		DeviceDiff func(*ttnpb.EndDevice)
 		Error      error
 	}{
 		{
-			Name: "adapted example from Semtech paper",
+			Name: "adapted example from Semtech paper/no rejections",
 			Device: &ttnpb.EndDevice{
+				FrequencyPlanID:   test.EUFrequencyPlanID,
+				LoRaWANPHYVersion: ttnpb.PHY_V1_0_2_REV_B,
 				MACState: &ttnpb.MACState{
 					CurrentParameters: ttnpb.MACParameters{
-						ADRDataRateIndex: 0,
-						ADRNbTrans:       0,
-						ADRTxPowerIndex:  1,
+						ADRNbTrans:      1,
+						ADRTxPowerIndex: 1,
+						Channels:        MakeDefaultEU868CurrentChannels(),
 					},
 					DesiredParameters: ttnpb.MACParameters{
-						ADRDataRateIndex: 5,
+						ADRDataRateIndex: ttnpb.DATA_RATE_4,
 						ADRNbTrans:       3,
 						ADRTxPowerIndex:  2,
+						Channels:         MakeDefaultEU868CurrentChannels(),
 					},
 				},
 				MACSettings: &ttnpb.MACSettings{
@@ -257,48 +294,173 @@ func TestAdaptDataRate(t *testing.T) {
 						Value: 2,
 					},
 				},
-				FrequencyPlanID: test.EUFrequencyPlanID,
-				RecentADRUplinks: adrMatrixToUplinks([]adrMatrixRow{
-					{FCnt: 10, MaxSNR: -6, GtwDiversity: 2},
-					{FCnt: 11, MaxSNR: -7, GtwDiversity: 2},
-					{FCnt: 12, MaxSNR: -25, GtwDiversity: 1},
-					{FCnt: 13, MaxSNR: -25, GtwDiversity: 1},
-					{FCnt: 14, MaxSNR: -10, GtwDiversity: 2},
-					{FCnt: 16, MaxSNR: -25, GtwDiversity: 1},
-					{FCnt: 17, MaxSNR: -10, GtwDiversity: 2},
-					{FCnt: 19, MaxSNR: -10, GtwDiversity: 3},
-					{FCnt: 20, MaxSNR: -6, GtwDiversity: 2},
-					{FCnt: 21, MaxSNR: -7, GtwDiversity: 2},
-					{FCnt: 22, MaxSNR: -25, GtwDiversity: 0},
-					{FCnt: 23, MaxSNR: -25, GtwDiversity: 1},
-					{FCnt: 24, MaxSNR: -10, GtwDiversity: 2},
-					{FCnt: 25, MaxSNR: -10, GtwDiversity: 2},
-					{FCnt: 26, MaxSNR: -25, GtwDiversity: 1},
-					{FCnt: 27, MaxSNR: -8, GtwDiversity: 2},
-					{FCnt: 28, MaxSNR: -10, GtwDiversity: 2},
-					{FCnt: 29, MaxSNR: -10, GtwDiversity: 3},
-					{FCnt: 30, MaxSNR: -9, GtwDiversity: 3},
-					{
-						FCnt: 31, MaxSNR: -7, GtwDiversity: 2,
-						TxSettings: ttnpb.TxSettings{
-							DataRate: ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_LoRa{
-									LoRa: &ttnpb.LoRaDataRate{
-										SpreadingFactor: 12,
-										Bandwidth:       125000,
-									},
-								},
-							},
-							DataRateIndex: 0,
-						},
-					},
-				}),
+				RecentADRUplinks: semtechPaperUplinks,
 			},
-			PHY: test.Must(band.All[band.EU_863_870].Version(ttnpb.PHY_V1_1_REV_B)).(band.Band),
 			DeviceDiff: func(dev *ttnpb.EndDevice) {
-				dev.MACState.DesiredParameters.ADRDataRateIndex = 4
-				dev.MACState.DesiredParameters.ADRNbTrans = 1
+				dev.MACState.DesiredParameters.ADRDataRateIndex = ttnpb.DATA_RATE_4
 				dev.MACState.DesiredParameters.ADRTxPowerIndex = 1
+				dev.MACState.DesiredParameters.ADRNbTrans = 1
+			},
+		},
+		{
+			Name: "adapted example from Semtech paper/rejected DR:(1,4)",
+			Device: &ttnpb.EndDevice{
+				FrequencyPlanID:   test.EUFrequencyPlanID,
+				LoRaWANPHYVersion: ttnpb.PHY_V1_0_2_REV_B,
+				MACState: &ttnpb.MACState{
+					CurrentParameters: ttnpb.MACParameters{
+						ADRNbTrans:      1,
+						ADRTxPowerIndex: 1,
+						Channels:        MakeDefaultEU868CurrentChannels(),
+					},
+					DesiredParameters: ttnpb.MACParameters{
+						Channels: MakeDefaultEU868CurrentChannels(),
+					},
+					RejectedADRDataRateIndexes: []ttnpb.DataRateIndex{
+						ttnpb.DATA_RATE_1, ttnpb.DATA_RATE_4,
+					},
+				},
+				MACSettings: &ttnpb.MACSettings{
+					ADRMargin: &pbtypes.FloatValue{
+						Value: 2,
+					},
+				},
+				RecentADRUplinks: semtechPaperUplinks,
+			},
+			DeviceDiff: func(dev *ttnpb.EndDevice) {
+				dev.MACState.DesiredParameters.ADRDataRateIndex = ttnpb.DATA_RATE_3
+				dev.MACState.DesiredParameters.ADRTxPowerIndex = 2
+				dev.MACState.DesiredParameters.ADRNbTrans = 1
+			},
+		},
+		{
+			Name: "adapted example from Semtech paper/rejected TXPower:(1)",
+			Device: &ttnpb.EndDevice{
+				FrequencyPlanID:   test.EUFrequencyPlanID,
+				LoRaWANPHYVersion: ttnpb.PHY_V1_0_2_REV_B,
+				MACState: &ttnpb.MACState{
+					CurrentParameters: ttnpb.MACParameters{
+						ADRNbTrans:      1,
+						ADRTxPowerIndex: 1,
+						Channels:        MakeDefaultEU868CurrentChannels(),
+					},
+					DesiredParameters: ttnpb.MACParameters{
+						Channels: MakeDefaultEU868CurrentChannels(),
+					},
+					RejectedADRTxPowerIndexes: []uint32{
+						1,
+					},
+				},
+				MACSettings: &ttnpb.MACSettings{
+					ADRMargin: &pbtypes.FloatValue{
+						Value: 2,
+					},
+				},
+				RecentADRUplinks: semtechPaperUplinks,
+			},
+			DeviceDiff: func(dev *ttnpb.EndDevice) {
+				dev.MACState.DesiredParameters.ADRDataRateIndex = ttnpb.DATA_RATE_4
+				dev.MACState.DesiredParameters.ADRTxPowerIndex = 0
+				dev.MACState.DesiredParameters.ADRNbTrans = 1
+			},
+		},
+		{
+			Name: "adapted example from Semtech paper/rejected TXPower:(0,1)",
+			Device: &ttnpb.EndDevice{
+				FrequencyPlanID:   test.EUFrequencyPlanID,
+				LoRaWANPHYVersion: ttnpb.PHY_V1_0_2_REV_B,
+				MACState: &ttnpb.MACState{
+					CurrentParameters: ttnpb.MACParameters{
+						ADRNbTrans:      1,
+						ADRTxPowerIndex: 1,
+						Channels:        MakeDefaultEU868CurrentChannels(),
+					},
+					DesiredParameters: ttnpb.MACParameters{
+						Channels: MakeDefaultEU868CurrentChannels(),
+					},
+					RejectedADRTxPowerIndexes: []uint32{
+						0, 1,
+					},
+				},
+				MACSettings: &ttnpb.MACSettings{
+					ADRMargin: &pbtypes.FloatValue{
+						Value: 2,
+					},
+				},
+				RecentADRUplinks: semtechPaperUplinks,
+			},
+			DeviceDiff: func(dev *ttnpb.EndDevice) {
+				dev.MACState.DesiredParameters.ADRDataRateIndex = ttnpb.DATA_RATE_3
+				dev.MACState.DesiredParameters.ADRTxPowerIndex = 2
+				dev.MACState.DesiredParameters.ADRNbTrans = 1
+			},
+		},
+		{
+			Name: "adapted example from Semtech paper/rejected DR:(1,4), rejected TXPower:(0,2,3)",
+			Device: &ttnpb.EndDevice{
+				FrequencyPlanID:   test.EUFrequencyPlanID,
+				LoRaWANPHYVersion: ttnpb.PHY_V1_0_2_REV_B,
+				MACState: &ttnpb.MACState{
+					CurrentParameters: ttnpb.MACParameters{
+						ADRNbTrans:      1,
+						ADRTxPowerIndex: 1,
+						Channels:        MakeDefaultEU868CurrentChannels(),
+					},
+					DesiredParameters: ttnpb.MACParameters{
+						Channels: MakeDefaultEU868CurrentChannels(),
+					},
+					RejectedADRTxPowerIndexes: []uint32{
+						0, 2, 3,
+					},
+					RejectedADRDataRateIndexes: []ttnpb.DataRateIndex{
+						ttnpb.DATA_RATE_1, ttnpb.DATA_RATE_4,
+					},
+				},
+				MACSettings: &ttnpb.MACSettings{
+					ADRMargin: &pbtypes.FloatValue{
+						Value: 2,
+					},
+				},
+				RecentADRUplinks: semtechPaperUplinks,
+			},
+			DeviceDiff: func(dev *ttnpb.EndDevice) {
+				dev.MACState.DesiredParameters.ADRDataRateIndex = ttnpb.DATA_RATE_3
+				dev.MACState.DesiredParameters.ADRTxPowerIndex = 1
+				dev.MACState.DesiredParameters.ADRNbTrans = 1
+			},
+		},
+		{
+			Name: "adapted example from Semtech paper/rejected DR:(3), rejected TXPower:(0,1)",
+			Device: &ttnpb.EndDevice{
+				FrequencyPlanID:   test.EUFrequencyPlanID,
+				LoRaWANPHYVersion: ttnpb.PHY_V1_0_2_REV_B,
+				MACState: &ttnpb.MACState{
+					CurrentParameters: ttnpb.MACParameters{
+						ADRNbTrans:      1,
+						ADRTxPowerIndex: 1,
+						Channels:        MakeDefaultEU868CurrentChannels(),
+					},
+					DesiredParameters: ttnpb.MACParameters{
+						Channels: MakeDefaultEU868CurrentChannels(),
+					},
+					RejectedADRTxPowerIndexes: []uint32{
+						0, 1,
+					},
+					RejectedADRDataRateIndexes: []ttnpb.DataRateIndex{
+						ttnpb.DATA_RATE_3,
+					},
+				},
+				MACSettings: &ttnpb.MACSettings{
+					ADRMargin: &pbtypes.FloatValue{
+						Value: 2,
+					},
+				},
+				RecentADRUplinks: semtechPaperUplinks,
+			},
+			DeviceDiff: func(dev *ttnpb.EndDevice) {
+				dev.MACState.DesiredParameters.ADRDataRateIndex = ttnpb.DATA_RATE_2
+				dev.MACState.DesiredParameters.ADRTxPowerIndex = 3
+				dev.MACState.DesiredParameters.ADRNbTrans = 1
 			},
 		},
 	} {
@@ -306,13 +468,15 @@ func TestAdaptDataRate(t *testing.T) {
 			a := assertions.New(t)
 
 			dev := CopyEndDevice(tc.Device)
-
-			err := adaptDataRate(dev, tc.PHY, ttnpb.MACSettings{})
-			if err != nil && !a.So(err, should.Equal, tc.Error) ||
-				err == nil && !a.So(err, should.BeNil) {
-				t.FailNow()
+			fp := FrequencyPlan(dev.FrequencyPlanID)
+			err := adaptDataRate(
+				log.NewContext(test.ContextWithT(test.Context(), t), test.GetLogger(t)),
+				dev,
+				Band(fp.BandID, dev.LoRaWANPHYVersion), ttnpb.MACSettings{},
+			)
+			if !a.So(err, should.Equal, tc.Error) {
+				t.Fatalf("ADR failed with: %s", err)
 			}
-
 			expected := CopyEndDevice(tc.Device)
 			if tc.DeviceDiff != nil {
 				tc.DeviceDiff(expected)