+++ pkg/networkserver/mac_link_adr_test.go
@@ -150,7 +150,7 @@ func TestNeedsLinkADRReq(t *testing.T) {
 			a := assertions.New(t)
 
 			dev := CopyEndDevice(tc.InputDevice)
-			res := deviceNeedsLinkADRReq(dev)
+			res := deviceNeedsLinkADRReq(dev, DefaultConfig.DefaultMACSettings.Parse())
 			if tc.Needs {
 				a.So(res, should.BeTrue)
 			} else {